@@ -1,30 +1,25 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { NavBarComponent } from './nav-bar/nav-bar.component';
-import { BackgroundImgComponent } from './background-img/background-img.component';
-import { AuthAppComponent } from './auth-app/auth-app.component';
-import { SecretComponent } from './secret/secret.component';
-import { ClientsComponent } from './clients/clients.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { LoginPageComponent } from './login-page/login-page.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MaterialModule } from './shared/material.module';
 import { RegisterPageComponent } from './register-page/register-page.component';
+import { HomeComponent } from './home/home.component';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    NavBarComponent,
-    BackgroundImgComponent,
-    AuthAppComponent,
-    SecretComponent,
-    ClientsComponent,
-    LoginPageComponent,
-    RegisterPageComponent,
+  declarations: [AppComponent, NavBarComponent, RegisterPageComponent, HomeComponent],
+  imports: [
+    FormsModule,
+    ReactiveFormsModule,
+    BrowserModule,
+    AppRoutingModule,
+    HttpClientModule,
+    MaterialModule,
   ],
-  imports: [BrowserModule, AppRoutingModule, HttpClientModule, BrowserAnimationsModule],
   providers: [],
   bootstrap: [AppComponent],
 })
