@@ -8,23 +8,26 @@ import { FormGroup, FormControl, Validators } from '@angular/forms';
   styleUrls: ['./register-page.component.css'],
 })
 export class RegisterPageComponent implements OnInit {
-  public registerForm!: FormGroup;
+  isLoginMode = true;
+  public registerForm: FormGroup;
 
   constructor(private authenticationService: AuthenticationService) {}
 
   ngOnInit() {
     this.registerForm = new FormGroup({
-      username: new FormControl('', Validators.required),
       email: new FormControl('', [Validators.required, Validators.email]),
       password: new FormControl('', Validators.required),
     });
   }
 
   public onSubmit() {
-    this.authenticationService.register(
-      this.registerForm.get('username')!.value,
-      this.registerForm.get('email')!.value,
-      this.registerForm!.get('password')!.value
-    );
+    const email = this.registerForm.get('email')!.value;
+    const password = this.registerForm!.get('password')!.value;
+    if (this.isLoginMode) {
+    } else {
+      this.authenticationService.signup(email, password).subscribe((res) => {
+        console.log(res);
+      });
+    }
   }
 }
