@@ -1,45 +1,28 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Router } from '@angular/router';
-import { AuthenticationClient } from '../clients/authentication.client';
+
+interface AuthResponseData {
+  idToken: string;
+  email: string;
+  refreshToken: string;
+  expiresIn: string;
+  localId: string;
+}
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthenticationService {
-  private tokenKey = 'token';
-
-  constructor(
-    private authenticationClient: AuthenticationClient,
-    private router: Router
-  ) {}
-
-  public login(username: string, password: string): void {
-    this.authenticationClient.login(username, password).subscribe((token) => {
-      localStorage.setItem(this.tokenKey, token);
-      this.router.navigate(['/']);
-    });
-  }
-
-  public register(username: string, email: string, password: string): void {
-    this.authenticationClient
-      .register(username, email, password)
-      .subscribe((token) => {
-        localStorage.setItem(this.tokenKey, token);
-        this.router.navigate(['/']);
-      });
-  }
-
-  public logout() {
-    localStorage.removeItem(this.tokenKey);
-    this.router.navigate(['/login']);
-  }
-
-  public isLoggedIn(): boolean {
-    let token = localStorage.getItem(this.tokenKey);
-    return token != null && token.length > 0;
-  }
-
-  public getToken(): string | null {
-    return this.isLoggedIn() ? localStorage.getItem(this.tokenKey) : null;
+  API_KEY = 'AIzaSyBxgWxb4RiFq0N5ap3mpexB5PSUKeZKjHw';
+  constructor(private http: HttpClient) {}
+  signup(email: string, password: string) {
+    return this.http.post<AuthResponseData>(
+      `https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=${this.API_KEY}`,
+      {
+        email: email,
+        password: password,
+        returnSecureToken: true,
+      }
+    );
   }
 }
