@@ -1,3 +1,13 @@
 export const environment = {
-  production: true
+  production: true,
+};
+const firebaseConfig = {
+  apiKey: 'AIzaSyBxgWxb4RiFq0N5ap3mpexB5PSUKeZKjHw',
+  authDomain: 'scuba-diving-3afc2.firebaseapp.com',
+  databaseURL: 'https://scuba-diving-3afc2-default-rtdb.firebaseio.com',
+  projectId: 'scuba-diving-3afc2',
+  storageBucket: 'scuba-diving-3afc2.appspot.com',
+  messagingSenderId: '554672536681',
+  appId: '1:554672536681:web:b5b5cb14b6641c096336f8',
+  measurementId: 'G-SSCTE3QMBY',
 };
